+++ src/components/atoms/Spinner/Spinner.tsx
@@ -1,10 +1,11 @@
 import React from 'react';
-import brandSpinner from './spinner-brand.gif';
-import negativeSpinner from './spinner-negative.gif';
-import actionSpinner from './spinner-action.gif';
 import deprecated from 'deprecated-prop-type';
+import brandSpinner from './spinner-brand.gif';
+import actionSpinner from './spinner-action.gif';
+import negativeSpinner from './spinner-negative.gif';
 import { useThemeContext } from '../../styles/Theme';
-import jlSpinner from '../../../content/images/jl-spinner.gif';
+import unbrandedSpinner from '../../../content/images/unbranded-spinner.gif';
+import unbrandedNegativeSpinner from '../../../content/images/unbranded-spinner-negative.gif';
 
 export type SpinnerStyling = 'primary' | 'secondary' | 'negative';
 
@@ -26,23 +27,26 @@ export interface SpinnerProps {
   negative?: boolean;
 }
 
-const animationData: Record<SpinnerStyling, string> = {
+const animationDataDefault: Record<SpinnerStyling, string> = {
   primary: brandSpinner,
   secondary: actionSpinner,
   negative: negativeSpinner,
 };
 
+const animationDataUnbranded: Record<SpinnerStyling, string> = {
+  primary: unbrandedSpinner,
+  secondary: unbrandedSpinner,
+  negative: unbrandedNegativeSpinner,
+};
+
 const Spinner: React.FC<SpinnerProps> = (props) => {
   const theme = useThemeContext();
 
+  const animationData = theme?.spinner?.spinnerTheme === 'unbranded' ? animationDataUnbranded : animationDataDefault;
+
   deprecated(props.negative, 'negative prop is deprecated, use styling="negative" instead');
 
-  const spinner =
-    theme.spinner.spinnerTheme === 'jl'
-      ? jlSpinner
-      : props.negative
-      ? negativeSpinner
-      : animationData[props.styling || 'primary'];
+  const spinner = props.negative ? negativeSpinner : animationData[props.styling || 'primary'];
 
   const size = props.size === 'small' ? '20px' : '40px';